Change level from enum to actual Table

+++ PhotonPiano.BusinessLogic/BusinessModel/Class/ClassDetailModel.cs
@@ -8,7 +8,7 @@ namespace PhotonPiano.BusinessLogic.BusinessModel.Class
 {
     public record ClassDetailModel : ClassModel
     {
-        public double PricePerSlots {get; set;}
+        public decimal PricePerSlots {get; set;}
         public int SlotsPerWeek {get; set; }
         public ICollection<StudentClassModel> StudentClasses { get; set; } = new List<StudentClassModel>();
         public ICollection<SlotModel> Slots { get; set; } = new List<SlotModel>();
